#21 fixed user merge error in makefile
