+++ include/Login.hpp
@@ -1,28 +1,14 @@
 #ifndef LOGIN_HPP
 #define LOGIN_HPP
 
+#include "User.hpp"
+
 #include <fstream>
 #include <string>
 #include <map>
 #include <memory>
 #include <iostream>
 
-class User {
-    public:
-        /*
-        * @brief takes in a username and password and account to create a new user
-        * @param std::string for username
-        * @param std::string for password
-        * @param std::string account_type
-        */
-        User(std::string, std::string, std::string);
-    private:
-        friend class Login;
-        std::string name;
-        std::string password;
-        std::string account_type;
-        int permission;
-};
 
 class Login {
     public:
